services/aes/requirements.txt
flask==3.0.0
sentence-transformers==2.7.0
scikit-learn==1.4.0
language-tool-python==2.7.1
torch==2.2.0
nltk==3.8.1
python-dotenv==1.0.0

# torch==2.2.0 --index-url https://download.pytorch.org/whl/cpu